refactor(capsule): split host_state.rs impl groups into #[path] siblings
host_state.rs was 988 lines, inside the danger zone of the 1000-line file-size CI cap. Extract three cohesive impl HostState method groups into sibling #[path] modules (effective_* overlay accessors, the recv-path invocation-context installer, and the per-connection principal binding accessors), dropping the file to 680 lines. Pure mechanical cut-and-paste: no signature, visibility, or behaviour change. The private debug_assert_invocation_field_set helper stays in host_state.rs (called from two of the moved groups via ancestor-private access). Doc-link super:: paths rewritten to absolute crate paths to survive the extra module nesting.

Closes #952.
