+++ CHANGELOG.md
@@ -55,6 +55,8 @@ Changelog tracking starts with 0.2.0. Prior versions were not tracked.
 
 - **Capsule-contributed CLI verbs (`astrid capsule <verb> [args...]`).** A capsule can now declare a top-level CLI verb by adding `kind = "cli"` to a `[[command]]` entry in its `Capsule.toml` (the default `kind = "slash"` is unchanged — existing manifests parse and behave identically). The CLI resolves the verb against the daemon's command registry, dispatches it to the providing capsule over IPC as a non-interactive one-shot (`cli.v1.command.run.<provider>` → `cli.v1.command.result.<req_id>`, 30s budget), and exits with the capsule's reported exit code. `astrid capsule run <provider> <verb>` disambiguates when two capsules provide the same verb. CLI-verb names are validated at manifest parse time (fail closed: `[a-z][a-z0-9-]*`, 1-32 chars, cannot shadow a built-in `astrid capsule` subcommand). The kernel routes but never interprets the run/result payloads — the contract is capsule-space. Install surfaces any new CLI verbs a capsule adds. Refs #891.
 
+- Split `host_state.rs` `impl HostState` method groups (the per-principal `effective_*` overlay accessors, the recv-path invocation-context installer, and the per-connection principal binding accessors) into `#[path]` sibling modules to stay under the 1000-line file-size cap; pure mechanical move, no behaviour change. Closes #952.
+
 ### Fixed
 
 - **Daemon stop/restart no longer orphans the `state.db` lock, and the MCP shim survives a daemon restart.** A crashed/wedged daemon left `astrid stop` deleting only the socket — the orphaned process kept the singleton lock, so the next `astrid start` died silently on "Database … is already locked". The daemon now records its PID **and canonical executable path** in `run/system.pid`; `stop`/`restart` resolve the live process's executable (`/proc/<pid>/exe` on Linux, `proc_pidpath` on macOS) and signal it (SIGTERM → grace → SIGKILL) **only when the PID is live and the exe matches** — guarding against PID reuse, so on any mismatch or inability to verify it fails secure (no signal; `restart` bails) rather than risk killing an unrelated recycled-PID process. Daemon boot stderr is now captured to `log/daemon-boot.log` instead of `/dev/null`, so a lock-acquisition failure is no longer silent. Separately, the `astrid mcp serve` shim now detects connection-loss on the response path (not only on send) and pre-heals before the next request; auto-retry is gated to idempotent requests (`tools/list`) — a `tools/call` is never auto-retried, since it may have executed. `run/system.pid` and `log/daemon-boot.log` are ephemeral (recreated each boot); no persistent-format change, no migration. Closes #939.
