+++ crates/astrid-capsule/src/engine/wasm/host_state.rs
@@ -589,18 +589,6 @@ impl HostState {
         ))
     }
 
-    /// Build a fresh, empty per-connection principal registry.
-    ///
-    /// Callers that construct a `HostState` outside the pooled-capsule path
-    /// (lifecycle hooks, the hook handler, tests) use this so they do not have
-    /// to name the `dashmap` type. The pooled path instead clones one shared
-    /// registry into every instance (issue #45/#852).
-    #[must_use]
-    pub fn new_connection_principals()
-    -> Arc<dashmap::DashMap<u32, astrid_core::principal::PrincipalId>> {
-        Arc::new(dashmap::DashMap::new())
-    }
-
     /// Default **async-I/O** host-call semaphore (host-derived, fd-clamped).
     ///
     /// Sized much larger than the blocking semaphore and clamped by the process
@@ -614,27 +602,6 @@ impl HostState {
     /// Set the inbound message sender.
     pub fn set_inbound_tx(&mut self, tx: mpsc::Sender<InboundMessage>) {
         self.inbound_tx = Some(tx);
-    }
-
-    /// Return the KV namespace for this capsule scoped to its principal.
-    ///
-    /// Format: `{principal}:capsule:{capsule_id}`. This is the same namespace
-    /// used when the `ScopedKvStore` was created, but exposed here for cases
-    /// where host functions need to construct the namespace dynamically.
-    #[must_use]
-    pub fn principal_kv_namespace(&self) -> String {
-        format!("{}:capsule:{}", self.principal, self.capsule_id)
-    }
-
-    /// Return the effective KV store for the current invocation.
-    ///
-    /// Uses `invocation_kv` if set (different principal), falls back to
-    /// the capsule's default `kv` store.
-    #[must_use]
-    pub fn effective_kv(&self) -> &ScopedKvStore {
-        #[cfg(debug_assertions)]
-        self.debug_assert_invocation_field_set(self.invocation_kv.is_some(), "invocation_kv");
-        self.invocation_kv.as_ref().unwrap_or(&self.kv)
     }
 
     /// Debug-only consistency check: when the caller's principal differs from
@@ -669,289 +636,14 @@ impl HostState {
             );
         }
     }
-
-    /// Return the effective home mount for the current invocation.
-    ///
-    /// Prefers `invocation_home` (set when serving a different principal)
-    /// over `home` (set at capsule load for the owning principal).
-    #[must_use]
-    pub fn effective_home(&self) -> Option<&PrincipalMount> {
-        self.invocation_home.as_ref().or(self.home.as_ref())
-    }
-
-    /// Return the effective tmp mount for the current invocation. Same
-    /// precedence as [`effective_home`](Self::effective_home).
-    #[must_use]
-    pub fn effective_tmp(&self) -> Option<&PrincipalMount> {
-        self.invocation_tmp.as_ref().or(self.tmp.as_ref())
-    }
-
-    /// Owned copy of the effective home root path.
-    ///
-    /// Convenience for host fs functions that need to pass the principal
-    /// home into a security-gate check running inside an `async move` block.
-    #[must_use]
-    pub fn effective_home_root_buf(&self) -> Option<PathBuf> {
-        self.effective_home().map(|m| m.root.clone())
-    }
-
-    /// Return the effective secret store for the current invocation.
-    ///
-    /// Prefers `invocation_secret_store` (set when serving a different
-    /// principal) over the load-time `secret_store`.
-    #[must_use]
-    pub fn effective_secret_store(&self) -> &Arc<dyn SecretStore> {
-        #[cfg(debug_assertions)]
-        self.debug_assert_invocation_field_set(
-            self.invocation_secret_store.is_some(),
-            "invocation_secret_store",
-        );
-        self.invocation_secret_store
-            .as_ref()
-            .unwrap_or(&self.secret_store)
-    }
-
-    /// Return the effective capsule log file for the current invocation.
-    ///
-    /// Same precedence as [`effective_secret_store`](Self::effective_secret_store).
-    /// Returns `None` if neither the invocation nor load-time log is open.
-    #[must_use]
-    pub fn effective_capsule_log(&self) -> Option<&Arc<std::sync::Mutex<std::fs::File>>> {
-        self.invocation_capsule_log
-            .as_ref()
-            .or(self.capsule_log.as_ref())
-    }
-
-    /// Return the principal whose budget should be charged for host-fn
-    /// side-effects in the current invocation.
-    ///
-    /// Prefers the invoking principal from [`caller_context`](Self::caller_context)
-    /// (set per-invocation by [`WasmEngine::invoke_interceptor`](super::WasmEngine::invoke_interceptor))
-    /// and falls back to the capsule owner's [`principal`](Self::principal) when
-    /// no caller is in scope — load-time host calls, tests, and daemons'
-    /// self-triggered paths run on the owner's budget, matching the VFS/KV
-    /// `effective_*` accessors.
-    #[must_use]
-    pub fn effective_principal(&self) -> astrid_core::principal::PrincipalId {
-        self.caller_context
-            .as_ref()
-            .and_then(|m| m.principal.as_deref())
-            .and_then(|p| astrid_core::principal::PrincipalId::new(p).ok())
-            .unwrap_or_else(|| self.principal.clone())
-    }
-
-    /// Return the effective quota profile for the current invocation.
-    ///
-    /// Prefers `invocation_profile` (set by
-    /// [`WasmEngine::invoke_interceptor`](super::WasmEngine::invoke_interceptor)
-    /// for the calling principal) and falls back to the process-global
-    /// [`PrincipalProfile::default_ref`](astrid_core::profile::PrincipalProfile::default_ref)
-    /// when no invocation profile is in scope — load-time host calls, tests,
-    /// and single-tenant deployments all legitimately run without one.
-    ///
-    /// The fallback path intentionally does **not** substitute the capsule
-    /// owner's profile: that would leak the owner's quotas to every
-    /// unauthenticated call path. Using `Default` preserves single-tenant
-    /// parity while keeping the security invariant honest.
-    #[must_use]
-    pub fn effective_profile(&self) -> &astrid_core::profile::PrincipalProfile {
-        match self.invocation_profile.as_deref() {
-            Some(p) => p,
-            None => astrid_core::profile::PrincipalProfile::default_ref(),
-        }
-    }
-
-    /// Install per-invocation context from an inbound IPC message picked
-    /// up via [`ipc::Host::ipc_recv`](crate::engine::wasm::host::ipc) /
-    /// [`ipc::Host::ipc_poll`].
-    ///
-    /// Mirrors the principal-isolation setup done in
-    /// [`WasmEngine::invoke_interceptor`](super::WasmEngine::invoke_interceptor)
-    /// for the dispatcher path, but driven by `recv`/`poll` so that
-    /// `run + ipc::recv` capsules (prompt-builder, registry,
-    /// context-engine) also stamp publishes with the publisher's
-    /// principal and route reads/writes to the invoking principal's
-    /// namespaces. Without this hook these capsules silently fall back
-    /// to the owner principal (`default` for the standard distro),
-    /// breaking chat for any non-default agent: the publish goes out
-    /// stamped `default`, downstream interceptors load the wrong KV
-    /// namespace, the turn-state phase doesn't match, and the chain
-    /// stalls.
-    ///
-    /// Sets up the subset relevant to publish stamping and per-principal
-    /// KV / log routing:
-    /// - [`caller_context`](Self::caller_context) — drives both
-    ///   [`effective_principal`](Self::effective_principal) and the
-    ///   `principal_str` chosen by `publish_inner`.
-    /// - [`invocation_kv`](Self::invocation_kv) — per-principal KV
-    ///   namespace; falls back to load-time `kv` on failure.
-    /// - [`invocation_capsule_log`](Self::invocation_capsule_log) —
-    ///   per-principal log file; falls back to load-time `capsule_log`
-    ///   when the principal has no home directory yet.
-    /// - [`invocation_profile`](Self::invocation_profile) — the publishing
-    ///   principal's quota profile (owner included), resolved through
-    ///   [`profile_cache`](Self::profile_cache) so per-principal ceilings
-    ///   (background-process count, IPC throughput, HTTP streams) apply on
-    ///   this path too; falls back to the process-global default on a missing
-    ///   cache or failed load.
-    ///
-    /// Skipped vs the interceptor path (each is independently
-    /// recoverable; documenting the gaps so the omissions are
-    /// auditable):
-    /// - `invocation_home` / `invocation_tmp` / `invocation_secret_store` —
-    ///   none of the current run+recv capsules touch home/tmp paths
-    ///   or secrets from the recv loop. Add when one starts to.
-    /// - `store_meter` — the per-invocation linear-memory ceiling stays the
-    ///   capsule owner's; the recv path does not re-target it per publisher
-    ///   the way `invoke_interceptor` does. Acceptable because the run+recv
-    ///   capsules are shared singletons whose per-call allocation is bounded
-    ///   by the bus message-size limits. Re-target when a recv-driven capsule
-    ///   needs per-principal memory enforcement.
-    pub(crate) fn install_recv_invocation_context(&mut self, msg: &astrid_events::ipc::IpcMessage) {
-        // Fast path: if the new message's principal matches whatever
-        // we already have installed, keep the existing
-        // `invocation_kv` / `invocation_capsule_log` rather than
-        // re-opening the namespace and log file. The chat-stack run
-        // loop calls this on every recv tick — re-init each time
-        // burns I/O and allocations for no behavioural change.
-        // An interceptor's caller is owned by the dispatch path
-        // (`WasmEngine::invoke_interceptor`), not by recv. Nested
-        // `ipc::recv` calls inside an interceptor must NOT overwrite
-        // it — otherwise a recv'd message from a different publisher
-        // (or the empty-batch clear path below) would silently flip
-        // every subsequent `publish_json` away from the principal the
-        // interceptor was dispatched under.
-        if self.interceptor_active {
-            return;
-        }
-
-        let new_principal = msg.principal.clone();
-        let existing_principal = self
-            .caller_context
-            .as_ref()
-            .and_then(|c| c.principal.clone());
-        if new_principal == existing_principal {
-            // Refresh the caller context so e.g. topic name / payload
-            // tracking stays current, but skip the expensive resets.
-            self.caller_context = Some(msg.clone());
-            return;
-        }
-
-        self.caller_context = Some(msg.clone());
-
-        // The publishing principal, parsed once. Used two different ways
-        // below, matching the split in the interceptor path
-        // (`invoke_interceptor`):
-        //
-        //   • QUOTA profile — resolved for EVERY publisher, the owner
-        //     included. `effective_profile()`'s fallback is the process-global
-        //     *default*, never the owner's profile, so an owner-published
-        //     message must still resolve the owner's profile or its on-disk
-        //     quotas are silently ignored. (For an owner with no profile file
-        //     the cache returns the default, so this is a no-op in the common
-        //     single-tenant case and only bites once an operator configures
-        //     the owner principal.)
-        //
-        //   • KV / log / env overrides — installed only when the publisher
-        //     DIFFERS from the load-time owner. The load-time `kv` /
-        //     `capsule_log` / `config` are already the owner's, so an
-        //     owner-published message has nothing to override and these are
-        //     cleared back to the load-time values.
-        let publisher: Option<astrid_core::PrincipalId> = msg
-            .principal
-            .as_deref()
-            .and_then(|p| astrid_core::PrincipalId::new(p).ok());
-
-        // Resolve the publisher's quota profile (owner included) so
-        // per-principal ceilings (background-process count, IPC throughput,
-        // HTTP streams) apply on the guest-pulled `recv` path too — not only
-        // the dispatcher-driven interceptor path. When `msg.principal` is
-        // absent/unparseable the owner's own profile is resolved, mirroring
-        // `invoke_interceptor`'s `owner_principal` fallback. Best-effort: a
-        // failed load logs and leaves `invocation_profile = None` (the same
-        // process-global default fall-back as a missing cache), never denying
-        // the message — the recv path has no error channel.
-        let profile_principal = publisher.clone().unwrap_or_else(|| self.principal.clone());
-        self.invocation_profile = self.profile_cache.as_ref().and_then(|cache| {
-            match cache.resolve(&profile_principal) {
-                Ok(profile) => Some(profile),
-                Err(e) => {
-                    tracing::warn!(
-                        principal = %profile_principal,
-                        error = %e,
-                        "recv-path profile resolve failed; per-principal quotas fall back to the default profile"
-                    );
-                    None
-                },
-            }
-        });
-
-        // KV / log / env scoping overrides only kick in for a non-owner
-        // publisher; an owner-published message clears them back to the
-        // load-time (owner) values.
-        let Some(p) = publisher.filter(|p| *p != self.principal) else {
-            self.invocation_kv = None;
-            self.invocation_capsule_log = None;
-            self.invocation_env_overlay = None;
-            return;
-        };
-
-        let ns = format!("{}:capsule:{}", p, self.capsule_id);
-        self.invocation_kv = match self.kv.with_namespace(&ns) {
-            Ok(kv) => Some(kv),
-            Err(e) => {
-                tracing::warn!(
-                    principal = %p,
-                    error = %e,
-                    "Failed to create invocation KV scope on ipc::recv path"
-                );
-                None
-            },
-        };
-
-        self.invocation_capsule_log = super::open_capsule_log(&p, self.capsule_id.as_str(), false);
-        self.invocation_env_overlay =
-            super::load_invocation_env_overlay(&p, self.capsule_id.as_str());
-    }
-
-    /// Bind `principal` to the connection identified by stream resource `rep`.
-    ///
-    /// Called by the `net.unix-listener.accept` path after a verified
-    /// per-connection principal challenge-response (issue #45/#852). Storage
-    /// only — see [`connection_principals`](Self::connection_principals).
-    pub(crate) fn bind_connection_principal(
-        &self,
-        rep: u32,
-        principal: astrid_core::principal::PrincipalId,
-    ) {
-        self.connection_principals.insert(rep, principal);
-    }
-
-    /// Return the verified principal bound to the connection identified by
-    /// stream resource `rep`, if any.
-    ///
-    /// The READ side of the registry (issue #45/#852). This change only
-    /// POPULATES the registry; the enforcement that consumes it — binding the
-    /// connection's verified principal to its outbound traffic — lands
-    /// separately, so the accessor is exercised by tests but not yet by any
-    /// non-test call path. `allow(dead_code)` keeps the accessor in place for
-    /// that follow-up without tripping `-D warnings` in the interim.
-    #[must_use]
-    #[allow(dead_code)]
-    pub(crate) fn connection_principal(
-        &self,
-        rep: u32,
-    ) -> Option<astrid_core::principal::PrincipalId> {
-        self.connection_principals.get(&rep).map(|e| e.clone())
-    }
-
-    /// Remove any verified-principal binding for the connection identified by
-    /// stream resource `rep`. Called when the stream resource drops so the
-    /// registry does not leak entries for closed connections.
-    pub(crate) fn unbind_connection_principal(&self, rep: u32) {
-        self.connection_principals.remove(&rep);
-    }
 }
+
+#[path = "host_state_connection.rs"]
+mod connection;
+#[path = "host_state_effective.rs"]
+mod effective;
+#[path = "host_state_invocation.rs"]
+mod invocation;
 
 impl std::fmt::Debug for HostState {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
